% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excludeCities.R
\name{add_geometry_portion}
\alias{add_geometry_portion}
\title{Add Geometry Portion}
\usage{
add_geometry_portion(sf_big, sf_small, match_col)
}
\arguments{
\item{sf_big}{sf the original sf}

\item{sf_small}{a clipped version of sf_big for which
you want to know the portion of each item that remains from
when it was sf_big}

\item{match_col}{string identifying the key column}
}
\value{
sf with a new column size_portion
}
\description{
Given a big geometry and a clipped subset, determine the
portion of the geometry in big represented in small.
Uses geometry type (line, polygon), to assess portion.
}
\examples{
\dontrun{
out_sf <- add_geometry_portion(sf_poly, out_sf, match_col = "someid")
}
}
