% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excludeCities.R
\name{exclude_cities}
\alias{exclude_cities}
\title{Exclude Cities from Geometry}
\usage{
exclude_cities(sf_poly, sf_cities, match_col)
}
\arguments{
\item{sf_poly}{simple features of any polygon}

\item{sf_cities}{simple features of cities}

\item{match_col}{string identifying the key column}
}
\description{
Remove cities in Kitsap County from the passed simple features.
The resulting sf can be used in st_intersection on other geometries.
It is actually a generic excluding function that returns sf_poly
without whatever intersects with sf_cities.
It does no geographic interpolation of variables; for example, if sf_poly
contained a population figure and was partially excluded, the value for
population would be exactly the same as the full polygon.
The portion of the geometry remaining is calculated.
}
\examples{
cities <- sf::read_sf(file.path("data","cities"))
outline <- sf::read_sf(file.path("data","outline"))
county_no_cities <- exclude_cities(outline, cities)
}
\keyword{cities}
