% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excludeCities.R
\name{excludeCitiesFromPolygon}
\alias{excludeCitiesFromPolygon}
\title{Exclude Cities from Polygon}
\usage{
excludeCitiesFromPolygon(sf_cities, sf_poly)
}
\arguments{
\item{sf_cities}{simple features of cities}

\item{sf_poly}{simple features of any polygon}
}
\description{
Remove cities in Kitsap County from the passed simple features.
The resulting sf can be used in st_intersection on other geometries.
}
\examples{
cities <- sf::read_sf(file.path("data","cities"))
outline <- sf::read_sf(file.path("data","outline"))
county_no_cities <- excludeCities(st_difference(outline, st_union(cities)))
}
\keyword{cities}
