% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKCData.R
\name{get_data_details}
\alias{get_data_details}
\title{Get Data Details}
\usage{
get_data_details(dname = "Kitsap_County_Outline")
}
\arguments{
\item{dataset_name}{string of valid data name
Can use \code{get_dataset_name()} to find it.
Or browse \verb{https://kitsap-od-kitcowa.hub.arcgis.com} until you find it.}
}
\value{
list of fields, type, and batches
}
\description{
Get Data Details
}
\examples{
get_data_details()
}
