% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKCData.R
\name{get_dataset_name}
\alias{get_dataset_name}
\title{Get Dataset Name}
\usage{
get_dataset_name(data_wanted = "outline")
}
\arguments{
\item{data_wanted}{string of what is desired
is not exhaustive
Defaults to outline}
}
\description{
Get Dataset Name
}
\examples{
get_dataset_name
}
