% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKCData.R
\name{get_kc_data}
\alias{get_kc_data}
\title{Get KC Data}
\usage{
get_kc_data(dataset_name = "Kitsap_County_Outline")
}
\arguments{
\item{dataset_name}{string of valid data name
Can use \code{get_dataset_name()} to find it.
Or browse \verb{https://kitsap-od-kitcowa.hub.arcgis.com} until you find it.}
}
\value{
sf or csv
}
\description{
Get KC Data
}
\examples{
outline <- get_kc_data("Kitsap_County_Outline")
}
